LayoutTests/imported/w3c/web-platform-tests/css/css-ui/historical/moz-user-modify-01-expected.txt


PASS -moz-user-modify is not supported
